+++ example/flutter_package_size_example.dart
@@ -1,6 +1,8 @@
 import 'package:flutter_package_size/flutter_package_size.dart';
+import 'package:flutter_package_size/src/app/cli.dart';
 
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
+class FlutterPackageSize {
+  Future<void> run(List<String> args) async {
+    await Cli().run(args);
+  }
 }