+++ lib/src/core/constants.dart
@@ -1,3 +1,3 @@
 const packageName = "flutter_package_size";
 
-const version = "0.1.0";
+const packageVersion = "0.1.0";