+++ lib/src/core/logger.dart
@@ -4,14 +4,20 @@ class Logger {
   }
 
   void success(String message) {
-    print("✓ $message");
+    print('✓ $message');
   }
 
   void warning(String message) {
-    print("⚠ $message");
+    print('⚠ $message');
   }
 
   void error(String message) {
-    print("✗ $message");
+    print('✗ $message');
+  }
+
+  void divider() {
+    print(
+      '--------------------------------------------------------',
+    );
   }
 }