+++ lib/src/data/datasources/build_result.dart
@@ -1,0 +1,5 @@
+class BuildResult {
+  final String analysisJsonPath;
+
+  BuildResult(this.analysisJsonPath);
+}